changed the apis a bit and added documentation.
