+++ public/app/features/dashboard/components/PanelEditor/usePanelLatestData.ts
@@ -1,4 +1,4 @@
-import { DataQueryError, LoadingState, PanelData, GetDataOptions } from '@grafana/data';
+import { DataQueryError, LoadingState, PanelData, QueryRunnerGetDataOptions } from '@grafana/data';
 import { useEffect, useRef, useState } from 'react';
 import { PanelModel } from '../../state';
 import { Unsubscribable } from 'rxjs';
@@ -13,7 +13,7 @@ interface UsePanelLatestData {
 /**
  * Subscribes and returns latest panel data from PanelQueryRunner
  */
-export const usePanelLatestData = (panel: PanelModel, options: GetDataOptions): UsePanelLatestData => {
+export const usePanelLatestData = (panel: PanelModel, options: QueryRunnerGetDataOptions): UsePanelLatestData => {
   const querySubscription = useRef<Unsubscribable>();
   const [latestData, setLatestData] = useState<PanelData>();
